Refactor: Reorganize frontend into auth, project, and core modules
- Moved files and folders to a new modular structure under `frontend/src/modules`.
- Updated all relevant import paths to reflect the new locations.
- This change aims to improve modularity and prepare for potential future micro-frontend architecture.
