+++ frontend/src/modules/auth/components/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom'
-import { useAuthStore } from '../../stores/authStore'
+import { useAuthStore } from '../stores/authStore'
 
 export default function ProtectedRoute({ children }) {
   const { user, loading } = useAuthStore()
