+++ frontend/src/modules/core/config/firebase.js
