+++ frontend/src/modules/core/main.jsx
