+++ frontend/src/modules/core/stores/themeStore.js
