+++ frontend/src/modules/project/components/CreateProjectForm.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuthStore } from "../../stores/authStore";
-import { useProjectStore } from "../../stores/projectStore";
+import { useAuthStore } from "../../../auth/stores/authStore";
+import { useProjectStore } from "../stores/projectStore";
 import { Calendar, FileText, FolderCheck, Target, Type } from "lucide-react";
 import { format } from "date-fns";
