+++ frontend/src/modules/project/components/Nodo.jsx
